requirements.txt
Django==5.1.2
djangorestframework==3.15.2
environs==11.0.0
payme-pkg==3.0.6
python-dotenv==1.0.1
requests==2.32.3
django-unfold==0.40.0